Remove binary wrapper jar and document build
