+++ .gitignore
@@ -1,3 +1,11 @@
-/build/
-/app/build/
+# Gradle files
+.gradle
+build/
+# Local configuration
+local.properties
+# Android Studio
+.idea
+*.iml
+
+# Gradle wrapper
 gradle/wrapper/gradle-wrapper.jar
