+++ app/proguard-rules.pro
@@ -1,0 +1,1 @@
+# ProGuard rules for the release build.
