+++ app/src/main/AndroidManifest.xml
@@ -1,19 +1,17 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android">
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="com.example.basic">
     <application
         android:allowBackup="true"
         android:label="@string/app_name"
         android:icon="@android:drawable/ic_menu_gallery"
         android:roundIcon="@android:drawable/ic_menu_gallery"
         android:supportsRtl="true"
-        android:theme="@style/Theme.RandomQr">
-        <activity
-            android:name=".MainActivity"
-            android:exported="true">
+        android:theme="@android:style/Theme.Material.Light">
+        <activity android:name=".MainActivity" android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
     </application>
-
 </manifest>
