--- app/src/main/res/values-night/themes.xml
@@ -1,3 +1,0 @@
-<resources xmlns:tools="http://schemas.android.com/tools">
-    <style name="Theme.RandomQr" parent="Theme.Material3.DayNight.NoActionBar" />
-</resources>
