+++ gradle.properties
@@ -1,9 +1,3 @@
-# This file was generated by the Gradle 'init' task.
-# https://docs.gradle.org/current/userguide/build_environment.html#sec:gradle_configuration_properties
-
-org.gradle.configuration-cache=true
-
-
+org.gradle.jvmargs=-Xmx1536m
 android.useAndroidX=true
-
 android.suppressUnsupportedCompileSdk=35
